+++ geo/src/lib.rs
@@ -67,7 +67,8 @@
 //!
 //! ## Boolean Operations
 //!
-//! - **[`BooleanOps`]**: combine or split (Multi)Polygons using intersecton, union, xor, or difference operations.
+//! - **[`BooleanOps`]**: Combine or split (Multi)Polygons using intersecton, union, xor, or difference operations
+//! - **[`UnaryUnion`]**: Efficient union of a collection of adjacent or overlapping [`Polygon`]s
 //!
 //! ## Outlier Detection
 //!